Code improvement

+++ app/src/main/java/com/irfankhoirul/popularmovie/modules/movie_detail/DetailMovieFragment.java
@@ -49,8 +49,20 @@ import static com.irfankhoirul.popularmovie.modules.movie_detail.DetailMoviePres
 import static com.irfankhoirul.popularmovie.util.ConstantUtil.BACKDROP_PATH_BASE_URL;
 import static com.irfankhoirul.popularmovie.util.ConstantUtil.POSTER_PATH_BASE_URL;
 
-/**
- * Created by Irfan Khoirul on 7/10/2017.
+/*
+ * Copyright 2017.  Irfan Khoirul Muhlishin
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 public class DetailMovieFragment extends Fragment implements DetailMovieContract.View {
@@ -358,7 +370,7 @@ public class DetailMovieFragment extends Fragment implements DetailMovieContract
         if (isTablet) {
             width = (int) (DisplayMetricUtils.getDeviceWidth(getActivity()) * 2.0f / 3.0f);
         } else {
-            width = (int) (DisplayMetricUtils.getDeviceWidth(getActivity()));
+            width = DisplayMetricUtils.getDeviceWidth(getActivity());
         }
         int height = (int) (9.0f / 16.0f * width);
         ViewGroup.LayoutParams backdropLayoutParams = ivMovieBackdrop.getLayoutParams();
